+++ spec/.mocharc.js
@@ -7,6 +7,7 @@ module.exports = {
 	autoIt: true,
 	reporter: "atom",
 	require: "chai/should",
+	recursive: true,
 	
 	beforeFinish(...args){
 		return 0; // Cheat the exit-status