+++ spec/filetypes/react/ignored.js
@@ -1,0 +1,7 @@
+"use strict";
+
+describe("Ignored files", () => {
+	it("should NOT run this", () => {
+		true.should.equal(false);
+	});
+});