+++ spec/filetypes/typescript-spec.ts
@@ -1,5 +1,5 @@
-describe("TypeScript specs", () => {
-	it("should run this test as well", async () => {
+describe("Filetype: TypeScript", () => {
+	it("runs TypeScript specs", async () => {
 		const foo:string = "Foo";
 		const bar:number = 45.24;
 		foo.should.be.a("string").that.equals("Foo");