+++ spec/subdir/another-spec.js
@@ -1,0 +1,5 @@
+describe("Subdirectory specs", () => {
+	it("should run this if --recursive is enabled", () => {
+		AtomMocha.options.recursive.should.equal(true);
+	});
+});